requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pymysql==1.1.0
cryptography==41.0.7
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0